share the dual vision tower between openvla-oft and vla-adapter
